Refactor Components Into Component Folder
To make the app more maintainable and easier to work with, the components need
to be moved out of the main App.js. The easiest solution for now is to create a
file for each component, and make that component the default export. As the app
grows and more components need to be added, refactoring into more functional specific
folders would make sense.
