+++ client/src/components/ActiveTagList.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import ActiveTag from './ActiveTag';
+
+class ActiveTagList extends React.Component {
+  render() {
+    const activeTags = this.props.tags.map(tag =>
+      <ActiveTag key={tag} tag={tag} handleTagDelete={this.props.handleTagDelete} />
+    );
+    return activeTags
+  }
+}
+
+export default ActiveTagList;
